+++ src/components/Attention/tests/Attention.test.tsx
@@ -1,0 +1,11 @@
+import React from "react";
+import { render, screen } from "@testing-library/react";
+import Attention from "../Attention";
+
+describe("Expand", () => {
+  test("Renders the ItemsPagination component", () => {
+    render(
+      <Attention title="Test" text="Text" date="2012-09-27" status="ACCEPTED" />
+    );
+  });
+});